ai-crawler/requirements.txt
beautifulsoup4
pyppeteer
fastapi[all]
uvicorn
python-dotenv
groq
boto3
datetime
requests
Pillow
sentencepiece
transformers[torch]
supabase==2.0.0
torch